Fix Results running Up until Laplacian cells
